Blank.mp4 (ffmpeg) w/AudioChannel @30FPS -1 hour
This file is set-up to be hosted from the github.io domain and serves as a media source that will allow Chrome to dedicate it to the compositor layer Across Tabs.
Therefore pinning the domain URL: [as example & default by script] " https://google.com/ ". 
With the TamperMonkey Script active on the Pinned Tab, this Video is streamed from Source here at git, and treated as validation for keeping FPS generated and synced to the Chrome Browser Compositor Layer Timing Element across all Tabs.  Ensuring that Lossless Scaling always has a source FPS to Generate from, which keeps the LS-FG Alive without the need for Constantly having to activate and deactivate at each playback.